{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = "{{cookiecutter.project_short_description}}"
authors = ["{{cookiecutter.author_name}} <{{cookiecutter.email}}>"]

[tool.poetry.dependencies]
python = "~{{cookiecutter.python_version}}"
fastapi-slim = "^0.115.0"
mangum = "^0.18.0"

[tool.poetry.group.dev.dependencies]
httpx = "^0.23.0"
uvicorn = "^0.30.6"

[tool.poetry.group.tests.dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
select = [
    "FAST", # FastAPI checks
    "I",    # isort
    "T20",  # No print statements
]

[tool.pycln]
all = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
